requirements.txt
pydicom
pytorch
numpy
pandas
matplotlib
opencv
scikit-image
scikit-learn
scipy
timm
torchvision